feat: close keep-alive connections if server is running out of fds
Add another exception class which will be thrown if accept() fails
due to lack of file descriptors. If that happens it'll now close
the connections which are in keep-alive state the longest, even
if they have not reached their time limit yet. Also modified the
getCurrentTime() function to return the time in milliseconds now.