fix: Fix segfaults with CGI, better error handling
Also implemented another singleton to keep track of all PIDs which
have been killed but where waitpid() was not successful yet, so we
don't end up having a lot of zombie processes at some point.

Currently with CGI it will always end up closing the connection,
because in sendResponse() we don't check if the current buffer is
empty, and then it throws an exception after send() returns 0.
The reason why I didn't add this check yet is, that since we don't
have Chunked encoding in CGI yet, curl would not receive the full
response anymore since without this header it would just wait for
EOF.