fix: Set CLOEXEC flag for most fds
The logfile will be closed in the child right after forking, to prevent the FD leaking into the executed process.
For ongoing uploads its more problematic, they are also using ofstream, which doesn't have a portable way to access the underlying fd.
If there is only one CGI at the same time its fine, but if there are multiple it would leak the fd from the other ongoing uploads into each child.
Maybe switch to open+write, or write a wrapper class similar to ofstream, which sets this flag automatically.