fix: Set CLOEXEC flag for upload_file_ and pipe
For the upload_file_ I created a FdWrap class, which just uses open + write
under the hood, and the destructor will close the fd automatically if it hasn't been closed before.
It doesn't have all the functions from std::ofstream but all the functions which we used
before have been reimplemented (with slight modifications, e.g. the write functions takes
a std::string instead of char * and size now).

Also setting the FD_CLOEXEC flag on the read end of the pipe in the parent, I overlooked that one before.